refactor: Complete conversion to pull-based metrics
- Update main.rs files directly (no new files)
- Remove all push gateway logic and dependencies
- Remove old push-based metrics files (metrics_exporter.rs, metrics.rs)
- Logger exposes metrics on port 9090 via HTTP
- Server exposes metrics on port 9091 via HTTP
- Prometheus scrapes directly from services

This is cleaner and simpler than push-based monitoring.
